only build targets when logfile does not exist
git-svn-id: https://svn.pengutronix.de/svn/ptxdist/trunks/ptxdist-0.10-trunk@4998 33e552b5-05e3-0310-8538-816dae2090ed

+++ scripts/dep_hunter.sh
@@ -15,8 +15,11 @@
 packages=$(ptxdist print PACKAGES | grep "PACKAGES is" | sed "s/PACKAGES is \"\(.*\)\"$/\1/")
 
 for i in $packages; do
-	echo $i > logfile;
-	ptxdist clean
-	ptxdist install $i
-	mv logfile logfile-$i
+	# only build if logfile does not exist to be able to restart the script
+	if [ ! -f logfile-$i ]; then
+		echo $i > logfile;
+		ptxdist clean
+		ptxdist install $i
+		mv logfile logfile-$i
+	fi
 done
